requirements.txt
amqp==2.6.1
asgiref==3.3.1
attrs==19.3.0
Automat==0.8.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
blinker==1.4
boto3==1.16.59
botocore==1.19.59
celery==4.4.7
certifi==2020.12.5
chardet==4.0.0
Click==7.0
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.1.5
django-anymail==8.1
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-environ==0.4.5
django-filebrowser==3.13.2
django-google-analytics-app==5.0.2
django-grappelli==2.14.3
django-heroku==0.3.1
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-mysql==4.3.0
django-reset-migrations==0.4.0
django-richtextfield==1.6
django-storages==1.11.1
entrypoints==0.3
filelock==3.4.0
gunicorn==20.0.4
httplib2==0.14.0
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jmespath==0.10.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kombu==4.6.11
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
mysqlclient==2.1.0
netifaces==0.10.4
oauthlib==3.1.0
pbr==5.8.0
pexpect==4.6.0
Pillow==8.1.0
platformdirs==2.4.0
psycopg2==2.8.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-decouple==3.4
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
requests-unixsocket==0.2.0
s3transfer==0.3.4
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
sos==4.1
soupsieve==2.2
sqlparse==0.4.1
ssh-import-id==5.10
structlog==21.1.0
systemd-python==234
testresources==2.0.1
Twisted==18.9.0
ubuntu-advantage-tools==27.4
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
vine==1.3.0
virtualenv==20.10.0
wadllib==1.3.3
whitenoise==5.2.0
zipp==1.0.0
zope.interface==4.7.1
